MCR-3483 rm singleton and use interface
